Move EVM related logic for helper into evm.rs
